ResourcesCaches

+++ server/src/main/java/lsfusion/server/physics/dev/integration/external/to/file/FileAlterationObserver.java
@@ -1,18 +1,21 @@
 package lsfusion.server.physics.dev.integration.external.to.file;
 
 import com.google.common.base.Throwables;
+import lsfusion.base.ResourceUtils;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.monitor.FileAlterationListenerAdaptor;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Paths;
+import java.util.Arrays;
 
 public class FileAlterationObserver extends org.apache.commons.io.monitor.FileAlterationObserver {
     private final FileChangesListener fileChangesListener;
-    public FileAlterationObserver(String src, String target) {
+    public FileAlterationObserver(String src, String target, String... extensions) {
         super(src);
-        this.fileChangesListener = new FileChangesListener(src, target);
+        this.fileChangesListener = new FileChangesListener(src, target, extensions);
         init();
     }
 
@@ -28,29 +31,35 @@ public class FileAlterationObserver extends org.apache.commons.io.monitor.FileAl
     private static class FileChangesListener extends FileAlterationListenerAdaptor {
         private final String src;
         private final String target;
-        public FileChangesListener(String src, String target) {
+        private final String[] extensions;
+        public FileChangesListener(String src, String target, String... extensions) {
             this.src = src;
             this.target = target;
+            this.extensions = extensions;
         }
 
         @Override
         public void onDirectoryCreate(File directory) {
-            getTargetFile(directory).mkdirs();
+            if (target != null)
+                getTargetFile(directory).mkdirs();
         }
 
         @Override
         public void onDirectoryDelete(File directory) {
-            try {
-                FileUtils.deleteDirectory(getTargetFile(directory));
-            } catch (IOException e) {
-                throw Throwables.propagate(e);
+            if (target != null) {
+                try {
+                    FileUtils.deleteDirectory(getTargetFile(directory));
+                } catch (IOException e) {
+                    throw Throwables.propagate(e);
+                }
             }
         }
 
         @Override
         public void onFileCreate(File file) {
             try {
-                FileUtils.copyFile(file, getTargetFile(file));
+                if (!clearCaches(file, true))
+                    FileUtils.copyFile(file, getTargetFile(file));
             } catch (IOException e) {
                 throw Throwables.propagate(e);
             }
@@ -59,7 +68,8 @@ public class FileAlterationObserver extends org.apache.commons.io.monitor.FileAl
         @Override
         public void onFileChange(File file) {
             try {
-                FileUtils.copyFile(file, getTargetFile(file));
+                if (!clearCaches(file, false))
+                    FileUtils.copyFile(file, getTargetFile(file));
             } catch (IOException e) {
                 throw Throwables.propagate(e);
             }
@@ -67,7 +77,17 @@ public class FileAlterationObserver extends org.apache.commons.io.monitor.FileAl
 
         @Override
         public void onFileDelete(File file) {
-            getTargetFile(file).delete();
+            if (!clearCaches(file, true))
+                getTargetFile(file).delete();
+        }
+
+        private boolean clearCaches(File file, boolean pathsChanged) {
+            if (extensions.length > 0) {
+                Arrays.stream(extensions).filter(extension -> FilenameUtils.getExtension(file.getName()).equals(extension))
+                        .forEach(extension -> ResourceUtils.clearResourceCaches(extension, pathsChanged, true));
+                return true;
+            }
+            return false;
         }
 
         private File getTargetFile(File file) {
